Made an upgrade activate after collecting 10 points
Upgrade slightly boost the power of melee attacks and speed

+++ Assets/Scripts/EnemyScripts/EnemyHealth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    public int maxEnemyHealth;
+    public float maxEnemyHealth;
     public bool isSmallEnemy;
     public bool isBigEnemy;
     public float hitTimer;
